refactor sql to async call
